+++ experiments/src/main/scala/it/unipd/dei/diversity/source/ChasmUniformRandomSpherePointSource.scala
@@ -21,7 +21,7 @@ import it.unipd.dei.diversity.Point
 import scala.util.Random
 
 class ChasmUniformRandomSpherePointSource(override val dim: Int,
-                                        override val n: Int,
+                                        override val n: Long,
                                         override val k: Int,
                                         override val distance: (Point, Point) => Double,
                                         val randomGen: Random)
